+++ previous-versions/accounting/pkg/models/operations/deletebillpayment.go
@@ -8,9 +8,12 @@ import (
 )
 
 type DeleteBillPaymentRequest struct {
+	// Unique identifier for a bill payment.
 	BillPaymentID string `pathParam:"style=simple,explode=false,name=billPaymentId"`
-	CompanyID     string `pathParam:"style=simple,explode=false,name=companyId"`
-	ConnectionID  string `pathParam:"style=simple,explode=false,name=connectionId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
+	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 }
 
 func (o *DeleteBillPaymentRequest) GetBillPaymentID() string {
@@ -35,13 +38,16 @@ func (o *DeleteBillPaymentRequest) GetConnectionID() string {
 }
 
 type DeleteBillPaymentResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
 	// OK
 	PushOperationSummary *shared.PushOperationSummary
-	StatusCode           int
-	RawResponse          *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *DeleteBillPaymentResponse) GetContentType() string {