+++ previous-versions/accounting/pkg/models/operations/getaccounttransaction.go
@@ -8,9 +8,12 @@ import (
 )
 
 type GetAccountTransactionRequest struct {
+	// Unique identifier for an account transaction.
 	AccountTransactionID string `pathParam:"style=simple,explode=false,name=accountTransactionId"`
-	CompanyID            string `pathParam:"style=simple,explode=false,name=companyId"`
-	ConnectionID         string `pathParam:"style=simple,explode=false,name=connectionId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
+	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 }
 
 func (o *GetAccountTransactionRequest) GetAccountTransactionID() string {
@@ -37,11 +40,14 @@ func (o *GetAccountTransactionRequest) GetConnectionID() string {
 type GetAccountTransactionResponse struct {
 	// Success
 	AccountTransaction *shared.AccountTransaction
-	ContentType        string
+	// HTTP response content type for this operation
+	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetAccountTransactionResponse) GetAccountTransaction() *shared.AccountTransaction {