+++ previous-versions/accounting/pkg/models/operations/getprofitandloss.go
@@ -8,12 +8,14 @@ import (
 )
 
 type GetProfitAndLossRequest struct {
+	// Unique identifier for a company.
 	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
 	// Number of months defining the period of interest.
 	PeriodLength int `queryParam:"style=form,explode=true,name=periodLength"`
 	// Number of periods with `periodLength` to compare.
-	PeriodsToCompare int     `queryParam:"style=form,explode=true,name=periodsToCompare"`
-	StartMonth       *string `queryParam:"style=form,explode=true,name=startMonth"`
+	PeriodsToCompare int `queryParam:"style=form,explode=true,name=periodsToCompare"`
+	// The month the report starts from.
+	StartMonth *string `queryParam:"style=form,explode=true,name=startMonth"`
 }
 
 func (o *GetProfitAndLossRequest) GetCompanyID() string {
@@ -45,13 +47,16 @@ func (o *GetProfitAndLossRequest) GetStartMonth() *string {
 }
 
 type GetProfitAndLossResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
 	// Success
 	ProfitAndLossReport *shared.ProfitAndLossReport1
-	StatusCode          int
-	RawResponse         *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetProfitAndLossResponse) GetContentType() string {